requirements.txt
astroid==2.2.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cx-Oracle==7.1.3
Flask==1.0.2
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
ipinfo==1.1.1
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
passlib==1.7.1
Pillow==6.0.0
pylint==2.3.1
requests==2.21.0
six==1.12.0
typed-ast==1.3.4
urllib3==1.24.2
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
